Reorganize chart-utils
