+++ packages/client/src/ui/portfolio/OasisPortfolio.tsx
@@ -12,8 +12,8 @@ import { i18nSelector } from "modules/settings/selectors";
 import React from "react";
 import { connect } from "react-redux";
 import { RouteComponentProps, withRouter } from "react-router";
-import { ChartData } from "tools/chart-utils";
 import { composeWithProps } from "tools/context-utils";
+import { ChartData } from "tools/cosmos-chart-utils";
 import { toDateKey } from "tools/date-utils";
 import { GraphQLGuardComponent } from "ui/GraphQLGuardComponents";
 import { DashboardError } from "ui/pages/DashboardPage";
